change to floats / timer implementation
timer is implemented in TJ Scene

+++ Assets/Scripts/FileLoading/Scenarios/Choices.cs
@@ -10,9 +10,9 @@ namespace FileLoading
         public string Choice;
         
         // How the choice changes the stats
-        public int Approval;
-        public int Efficiency;
-        public int Environment;
-        public int Finance;
+        public float Approval;
+        public float Efficiency;
+        public float Environment;
+        public float Finance;
     }
 }
